+++ lib/domain/interfaces/i_audio_provider.dart
@@ -1,6 +1,6 @@
 import '../entities/audio_stream.dart';
-import '../entities/aura_state_enum.dart';
 
 abstract class IAudioProvider {
-  Future<List<AudioStream>> fetchStreams(AuraState state);
+  // Artık sadece durum değil, öğrenilmiş spesifik "tag" (tür) aranıyor
+  Future<List<AudioStream>> fetchStreams(String tag);
 }